Extract LogCacheBundle class
