+++ src/app/bug-reporting/environment/renderer-environment-collector.js
@@ -17,8 +17,9 @@
 
 // @flow
 
-import type { EnvironmentCollector, SerializedLogCaches } from './environment-collector'
+import type { EnvironmentCollector } from './environment-collector'
 import type { SyncRendererCommunication } from '../../communication/sync/sync-communication'
+import type { SerializedLogCaches } from '../../logging/log-cache-bundle'
 
 class RendererEnvironmentCollector implements EnvironmentCollector {
   _mysterionReleaseId: string
