+++ src/app/communication/sync/sync-renderer-communication.js
@@ -21,7 +21,7 @@ import logger from '../../logger'
 import messages from '../messages'
 import type { SyncSender } from './sync'
 import type { SyncRendererCommunication } from './sync-communication'
-import type { SerializedLogCaches } from '../../bug-reporting/environment/environment-collector'
+import type { SerializedLogCaches } from '../../logging/log-cache-bundle'
 import type { LogDTO } from '../dto'
 
 /**
